requirements.txt
beautifulsoup4==4.12.3
cachetools==5.3.2
caldav==1.3.9
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
filelock==3.13.1
google-api-core==2.17.0
google-api-python-client==2.117.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
html5lib==1.1
httplib2==0.22.0
icalendar==5.0.11
identify==2.5.34
idna==3.6
lxml==5.1.0
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.0
platformdirs==4.2.0
pre-commit==3.6.1
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
recurring-ical-events==2.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.0
virtualenv==20.25.0
vobject==0.9.6.1
webencodings==0.5.1
x-wr-timezone==0.0.6
